+++ xs/include/xs/syntax_ast.h
@@ -162,6 +162,7 @@ enum
   XS_SYNTAX_FLAG_OPERATOR = 1U << 14,
   XS_SYNTAX_FLAG_OVERLOAD = 1U << 15,
   XS_SYNTAX_FLAG_INNER_ATTRIBUTE = 1U << 16,
+  XS_SYNTAX_FLAG_EXTERN = 1U << 17,
 };
 
 typedef struct XsSyntaxNode XsSyntaxNode;